refactor: 💡 Refactor default option in scope field and desc (#328)
* refactor: 💡 Refactor default option in scope field and desc

Refactor default option for scope field to be 'Choose an option' instead
of 'Global', so user can explicitly select their scope.  Refactor scope
field descriptions to give different context based on if the selection
is 'global' or an 'org' scope.

✅ Closes: ICU-9641

* refactor: 💡 refactor translations

Added correct translations for default, global, and org. Added small bit
of logic to help with switch of text on `onChange` event.