feat: 🎸 distinguish sorting by model attributes vs json api key
By default store.query({ sort }) will expect the sort object to have an
attribute specified based on attributes defined by the model. The
sorting will actually happen on "keys" in the json api formatted
"attributes" object, which might contain keys that are not defined as
attr() on the model. This is true for the current default sort of
created_time. Making this distinction will help avoid querying on
insider json api attributes from the payload without a refactor or
some explicit handling. In the case of "created_time", it is identified
as a "universally sortable key" for all models in the code but
currently is not exposed as an option that can be passed via { sort }
since it's the current default.