refactor: 💡 use schema instead of property for getSortableValue
Use the schema for a stricter lookup of existing attribute. The two
value exceptions currently supported are created_time and id.
created_time exists on every record from the api but may not be defined
on the model. id exists on every record but is not considered an
attribute by json api record definition