refactor: 💡 removed aria-label from button set
This was not sending the aria-label down to the Edit Users detail button.
However, taking a look further at Hds::Button, this component
does not allow aria-label to be set unless it is an icon only button.
With that said, this cleans up the translation structure and removes the
aria-label translation and mark-up.

✅ Closes: FE-112